Fix compatibility with the latest Sanic release
